refactor(skills): extract shared sync helpers, fix TUI dead branch and HttpClient leak
- Extract SkillSyncHelpers with shared ComputeSha256, ValidateResourcePath,
  ReadSyncState/WriteSyncState, ReplaceSkillDirectoryAsync, and
  DownloadedSkillFile — eliminates duplication between SystemSkillSyncService
  and ServerFeedSkillSyncService
- Cache JsonSerializerOptions as static readonly instead of allocating per write
- Create one HttpClient per feed instead of per-download in
  ServerFeedSkillSyncService.DownloadAndVerifyAsync
- Remove dead branch in SkillFeedsStepView.BuildConnectPrompt where both
  if/else paths called AdvanceStep()
- Remove redundant _lastProbeSkillCount >= 0 check in ProbeSucceeded (int is
  always >= 0)
